HARP-9443: Implement the GeoJson datasource using omv-datasource.
This change removes the GeoJson specific parsing and decoders
of the module @here/harp-geojson-datasource.

Support for GeoJson is now provided by @here/harp-omv-datasource
and the @here/harp-geojson-datasource simply provides a
`GeoJsonDataSource` class that extends `OmvDataSource` for backward
compatibility.
