+++ @here/harp-examples/decoder/decoder.ts
@@ -10,13 +10,11 @@ declare let self: Worker & {
 
 self.importScripts("three.min.js");
 
-import { GeoJsonTileDecoderService } from "@here/harp-geojson-datasource/index-worker";
 import { OmvTileDecoderService, OmvTilerService } from "@here/harp-omv-datasource/index-worker";
 import { CustomDecoderService } from "./custom_decoder";
 
 OmvTileDecoderService.start();
 OmvTilerService.start();
-GeoJsonTileDecoderService.start();
 
 //Following code is only needed for datasource_custom example.
 // snippet:custom_datasource_example_custom_decoder_service_start.ts
