+++ @here/harp-examples/src/datasource_geojson_choropleth.ts
@@ -5,11 +5,15 @@
  */
 
 import { StyleDeclaration, StyleSet, Theme } from "@here/harp-datasource-protocol";
-import { GeoJsonDataProvider } from "@here/harp-geojson-datasource";
 import { GeoCoordinates } from "@here/harp-geoutils";
 import { MapControls, MapControlsUI } from "@here/harp-map-controls";
 import { MapView } from "@here/harp-mapview";
-import { APIFormat, AuthenticationMethod, OmvDataSource } from "@here/harp-omv-datasource";
+import {
+    APIFormat,
+    AuthenticationMethod,
+    GeoJsonDataProvider,
+    OmvDataSource
+} from "@here/harp-omv-datasource";
 import * as THREE from "three";
 import { apikey, copyrightInfo } from "../config";
