+++ @here/harp-geojson-datasource/index-worker.ts
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from "./lib/GeoJsonDecoder";
+export * from "./lib/GeoJsonTileDecoderService";
