+++ @here/harp-geojson-datasource/package.json
@@ -24,22 +24,12 @@
     },
     "license": "Apache-2.0",
     "dependencies": {
-        "@here/harp-datasource-protocol": "^0.14.0",
-        "@here/harp-fetch": "^0.14.0",
-        "@here/harp-geometry": "^0.14.0",
-        "@here/harp-geoutils": "^0.14.0",
-        "@here/harp-lines": "^0.14.0",
-        "@here/harp-mapview": "^0.14.0",
-        "@here/harp-mapview-decoder": "^0.14.0",
-        "@here/harp-text-canvas": "^0.14.0",
-        "@here/harp-transfer-manager": "^0.14.0",
-        "@here/harp-utils": "^0.14.0",
-        "earcut": "^2.1.3"
+        "@here/harp-omv-datasource": "^0.14.0",
+        "@here/harp-utils": "^0.14.0"
     },
     "devDependencies": {
         "@here/harp-test-utils": "^0.14.0",
         "@types/chai": "^4.1.2",
-        "@types/earcut": "^2.1.0",
         "@types/mocha": "^5.2.7",
         "@types/node": "^12.0.8",
         "chai": "^4.0.2",
