+++ @here/harp-omv-datasource/index.ts
@@ -10,3 +10,4 @@ export * from "./lib/OmvDecoderDefs";
 export * from "./lib/OmvTile";
 export * from "./lib/OmvDebugLabelsTile";
 export * from "./lib/OmvRestClient";
+export * from "./lib/GeoJsonDataProvider";
